+++ cloud/acp-admin-cloud-dependencies/src/main/kotlin/pers/acp/admin/common/event/ExecuteBusEvent.kt
@@ -6,4 +6,4 @@ import org.springframework.cloud.bus.event.RemoteApplicationEvent
  * @author zhang by 19/03/2019
  * @since JDK 11
  */
-class ExecuteBusEvent(origin: String?, destination: String?, val executeType: String, val paramList: List<String>, source: Any = Object()) : RemoteApplicationEvent(source, origin, destination)
+class ExecuteBusEvent(origin: String?, destination: String?, val message: String, val paramList: List<String>, source: Any = Object()) : RemoteApplicationEvent(source, origin, destination)